most optimization cython

+++ sockets/binance_socket_c.pyx
@@ -156,8 +156,8 @@ cdef printer(asks, bids, symbol):
     Function to process the received messages
     param msg: input message
     """
-    cdef asks_price = asks[:,0].astype(np.float16)
-    cdef asks_quantity = asks[:,1].astype(np.float16)
+    cdef asks_price = asks[:,0].astype(np.float32)
+    cdef asks_quantity = asks[:,1].astype(np.float32)
     cdef user_max_amount = float(db.get_info_col('max_amount'))
     cdef quantity = 0
     cdef int count = 0
@@ -173,8 +173,8 @@ cdef printer(asks, bids, symbol):
     cdef asks_amount = quantity
     cdef asks_avg_price = mean_price/count
 
-    cdef bids_price = bids[:,0].astype(np.float16)
-    cdef bids_quantity = bids[:,1].astype(np.float16)
+    cdef bids_price = bids[:,0].astype(np.float32)
+    cdef bids_quantity = bids[:,1].astype(np.float32)
 
     for i, val in enumerate(bids_quantity):
         if usdt_quantity < user_max_amount:
